Move some roundtrip tests to common crate
