+++ janus_server/src/message.rs
@@ -647,9 +647,7 @@ pub mod test_util {
 mod tests {
     use super::*;
     use assert_matches::assert_matches;
-    use janus::message::{
-        Duration, ExtensionType, HpkeAeadId, HpkeConfigId, HpkeKdfId, HpkeKemId, Time,
-    };
+    use janus::message::{Duration, ExtensionType, HpkeConfigId, Time};
     use lazy_static::lazy_static;
 
     lazy_static! {
@@ -860,53 +858,6 @@ mod tests {
                     "0000000000000000", // duration
                 ),
             ),
-        ])
-    }
-
-    #[test]
-    fn roundtrip_task_id() {
-        roundtrip_encoding(&[
-            (
-                TaskId::new([u8::MIN; 32]),
-                "0000000000000000000000000000000000000000000000000000000000000000",
-            ),
-            (
-                TaskId::new([
-                    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
-                    22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
-                ]),
-                "000102030405060708090A0B0C0D0E0F101112131415161718191A1B1C1D1E1F",
-            ),
-            (
-                TaskId::new([u8::MAX; 32]),
-                "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF",
-            ),
-        ])
-    }
-
-    #[test]
-    fn roundtrip_hpke_kem_id() {
-        roundtrip_encoding(&[
-            (HpkeKemId::P256HkdfSha256, "0010"),
-            (HpkeKemId::X25519HkdfSha256, "0020"),
-        ])
-    }
-
-    #[test]
-    fn roundtrip_hpke_kdf_id() {
-        roundtrip_encoding(&[
-            (HpkeKdfId::HkdfSha256, "0001"),
-            (HpkeKdfId::HkdfSha384, "0002"),
-            (HpkeKdfId::HkdfSha512, "0003"),
-        ])
-    }
-
-    #[test]
-    fn roundtrip_hpke_aead_id() {
-        roundtrip_encoding(&[
-            (HpkeAeadId::Aes128Gcm, "0001"),
-            (HpkeAeadId::Aes256Gcm, "0002"),
-            (HpkeAeadId::ChaCha20Poly1305, "0003"),
         ])
     }
